Refactor stdio build
